# swagger/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API documentation",
        "title": "Go Boilerplate API",
        "contact": {
            "name": "Abdan Hafidz",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/academy/{academy_slug}/exam/{exam_slug}/attempt": {
            "get": {
                "description": "Start an attempt for a specific exam in an academy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academy Exam"
                ],
                "summary": "Attempt Academy Exam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Academy Slug",
                        "name": "academy_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exam Slug",
                        "name": "exam_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-models_ExamAcademyAttempt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/authentication/external-login": {
            "post": {
                "description": "Authenticate user using external OAuth provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "External Authentication",
                "parameters": [
                    {
                        "description": "External Auth Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ExternalAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-dto_AuthenticatedUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/authentication/login": {
            "post": {
                "description": "Authenticate user and obtain access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Sign In Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-dto_AuthenticatedUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/authentication/register": {
            "post": {
                "description": "Register a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Registration",
                "parameters": [
                    {
                        "description": "Sign Up Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse-models_Account"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AuthenticatedUser": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/models.Account"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.ExternalAuthRequest": {
            "type": "object",
            "required": [
                "oauth_id",
                "oauth_provider"
            ],
            "properties": {
                "oauth_id": {
                    "type": "string"
                },
                "oauth_provider": {
                    "type": "string"
                }
            }
        },
...
        "dto.SignInRequest": {
            "type": "object",
            "required": [
                "email_or_username",
                "password"
            ],
            "properties": {
                "email_or_username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.SignUpRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-dto_AuthenticatedUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AuthenticatedUser"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-models_Account": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Account"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.SuccessResponse-models_ExamAcademyAttempt": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.ExamAcademyAttempt"
                },
                "message": {},
                "meta_data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Academy": {
            "type": "object",
            "properties": {
                "academy_progress": {
                    "$ref": "#/definitions/models.AcademyProgress"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AcademyMaterial"
                    }
                },
                "materials_count": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "register_status": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.AcademyContent": {
            "type": "object",
            "properties": {
                "academy_content_progress": {
                    "$ref": "#/definitions/models.AcademyContentProgress"
                },
                "contents": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "material_id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.AcademyContentProgress": {
            "type": "object",
            "properties": {
                "academy_id": {
                    "type": "string"
                },
                "account_id": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "content_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.AcademyMaterial": {
            "type": "object",
            "properties": {
                "academy_id": {
                    "type": "string"
                },
                "academy_material_progress": {
                    "$ref": "#/definitions/models.AcademyMaterialProgress"
                },
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AcademyContent"
                    }
                },
                "contents_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.AcademyMaterialProgress": {
            "type": "object",
            "properties": {
                "academy_id": {
                    "type": "string"
                },
                "account_id": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material_id": {
                    "type": "string"
                },
                "progress": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "total_completed_contents": {
                    "type": "integer"
                }
            }
        },
...
        "models.AcademyProgress": {
            "type": "object",
            "properties": {
                "academy_id": {
                    "type": "string"
                },
                "account_id": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "progress": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "total_completed_materials": {
                    "type": "integer"
                }
            }
        },
        "models.Account": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_detail_completed": {
                    "type": "boolean"
                },
                "is_email_verified": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.Exam": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/models.ExamConfiguration"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id_exam": {
                    "type": "string"
                },
                "proctoring": {
                    "$ref": "#/definitions/models.ExamProctoring"
                },
                "randomize": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ExamAcademyAnswer": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "exam_attempt": {
                    "$ref": "#/definitions/models.ExamAcademyAttempt"
                },
                "id": {
                    "type": "string"
                },
                "id_attempt": {
                    "type": "string"
                },
                "id_question": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ExamAcademyAttempt": {
            "type": "object",
            "properties": {
                "academy": {
                    "$ref": "#/definitions/models.Academy"
                },
                "account": {
                    "$ref": "#/definitions/models.Account"
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExamAcademyAnswer"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "due_at": {
                    "type": "string"
                },
                "exam": {
                    "$ref": "#/definitions/models.Exam"
                },
                "id_academy": {
                    "type": "string"
                },
                "id_account": {
                    "type": "string"
                },
                "id_attempt": {
                    "type": "string"
                },
                "id_exam": {
                    "type": "string"
                },
                "mark": {
                    "type": "number"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Questions"
                    }
                },
                "remaining_time": {
                    "type": "integer"
                },
                "submitted": {
                    "type": "boolean"
                }
            }
        },
        "models.ExamConfiguration": {
            "type": "object",
            "properties": {
                "allow_retake": {
                    "type": "boolean"
                },
                "allow_review": {
                    "type": "boolean"
                },
                "enable_timer": {
                    "type": "boolean"
                },
                "id_exam": {
                    "type": "string"
                },
                "id_result": {
                    "type": "string"
                }
            }
        },
...
        "models.ExamProctoring": {
            "type": "object",
            "properties": {
                "disable_copy_paste": {
                    "type": "boolean"
                },
                "enable_exam_browser": {
                    "type": "boolean"
                },
                "enable_eye_tracking": {
                    "type": "boolean"
                },
                "enable_full_screen": {
                    "type": "boolean"
                },
                "enable_tab_block": {
                    "type": "boolean"
                },
                "enable_vad": {
                    "type": "boolean"
                },
                "enable_webcam": {
                    "type": "boolean"
                },
                "id_exam": {
                    "type": "string"
                },
                "id_result": {
                    "type": "string"
                }
            }
        },
...
        "models.ProblemSet": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id_problem_set": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Questions": {
            "type": "object",
            "properties": {
                "ans_key": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "corr_mark": {
                    "type": "number"
                },
                "id_problem_set": {
                    "type": "string"
                },
                "id_question": {
                    "type": "string"
                },
                "incorr_mark": {
                    "type": "number"
                },
                "null_mark": {
                    "type": "number"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "problem_set": {
                    "$ref": "#/definitions/models.ProblemSet"
                },
                "question": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}